suppress stringop-trunction warning
Currently the following stringop-truncation warning is generated with
gcc 11.2.1:

In file included from ucurr.cpp:25:
In constructor ‘CReg::CReg(const UChar*, const char*)’,
    inlined from ‘static const void* CReg::reg(const UChar*,
                                               const char*,
                                               UErrorCode*)’
    at ucurr.cpp:397:41,
    inlined from ‘const void* ucurr_register_70(const UChar*,
                                                const char*,
                                                UErrorCode*)’
    at ucurr.cpp:469:25:
cstring.h:43:70: warning:
‘char* strncpy(char*, const char*, size_t)’ output truncated before
terminating nul copying as many bytes from a string as its length
[-Wstringop-truncation]
   43 | #define uprv_strncpy(dst, src, size)
            U_STANDARD_CPP_NAMESPACE strncpy(dst, src, size)
ucurr.cpp:387:9: note: in expansion of macro ‘uprv_strncpy’
  387 |         uprv_strncpy(id, _id, len);
      |         ^~~~~~~~~~~~
ucurr.cpp: In function ‘const void* ucurr_register_70(const UChar*,
                                                      const char*,
                                                      UErrorCode*)’:
cstring.h:37:57: note: length computed here
   37 | #define uprv_strlen(str) U_STANDARD_CPP_NAMESPACE strlen(str)
ucurr.cpp:381:32: note: in expansion of macro ‘uprv_strlen’
  381 |         int32_t len = (int32_t)uprv_strlen(_id);
      |                                ^~~~~~~~~~~

This commit adds a GCC ignore for this warnings.

The motivation for ignoring this warning is that uprv_strlen is called
to calculate the number of bytes to copy later in the call to
uprv_strncpy. uprv_strlen uses strlen which calculates the length of
the string passed in, excluding the termination null byte, and it looks
like this is the intention of the code, to not include the termination
null byte, hence it should be alright to ignore this warning (if these
assumptions are correct that is).